Create README.md
This repository is an application designated to help with studying for Microsoft Windows Networking Essentials. It takes all of the questions from the end of the Chapters to make it easier to study. Honestly though, this is a very sloppy application and is intended to demonstrate how to use Android Studio and using Git + GitHub together. It worked, and here it is for anyone to check out!